(improvment)(chat) refactor the jsqlparser code
