+++ chat/core/src/main/java/com/tencent/supersonic/chat/corrector/GlobalAfterCorrector.java
@@ -1,8 +1,9 @@
 package com.tencent.supersonic.chat.corrector;
 
 import com.tencent.supersonic.chat.api.pojo.SemanticCorrectInfo;
+import com.tencent.supersonic.common.util.jsqlparser.SqlParserAddHelper;
+import com.tencent.supersonic.common.util.jsqlparser.SqlParserSelectFunctionHelper;
 import com.tencent.supersonic.common.util.jsqlparser.SqlParserSelectHelper;
-import com.tencent.supersonic.common.util.jsqlparser.SqlParserUpdateHelper;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.expression.Expression;
@@ -15,12 +16,12 @@ public class GlobalAfterCorrector extends BaseSemanticCorrector {
 
         super.correct(semanticCorrectInfo);
         String sql = semanticCorrectInfo.getSql();
-        if (!SqlParserSelectHelper.hasAggregateFunction(sql)) {
+        if (!SqlParserSelectFunctionHelper.hasAggregateFunction(sql)) {
             return;
         }
         Expression havingExpression = SqlParserSelectHelper.getHavingExpression(sql);
         if (Objects.nonNull(havingExpression)) {
-            String replaceSql = SqlParserUpdateHelper.addFunctionToSelect(sql, havingExpression);
+            String replaceSql = SqlParserAddHelper.addFunctionToSelect(sql, havingExpression);
             semanticCorrectInfo.setSql(replaceSql);
         }
         return;
