+++ common/src/main/java/com/tencent/supersonic/common/util/jsqlparser/SqlParserSelectHelper.java
@@ -8,7 +8,11 @@ import java.util.Set;
 import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.JSQLParserException;
 import net.sf.jsqlparser.expression.Expression;
+import net.sf.jsqlparser.expression.ExpressionVisitorAdapter;
 import net.sf.jsqlparser.expression.Function;
+import net.sf.jsqlparser.expression.operators.conditional.AndExpression;
+import net.sf.jsqlparser.expression.operators.conditional.OrExpression;
+import net.sf.jsqlparser.expression.operators.conditional.XorExpression;
 import net.sf.jsqlparser.expression.operators.relational.ComparisonOperator;
 import net.sf.jsqlparser.parser.CCJSqlParserUtil;
 import net.sf.jsqlparser.schema.Column;
@@ -21,6 +25,8 @@ import net.sf.jsqlparser.statement.select.Select;
 import net.sf.jsqlparser.statement.select.SelectBody;
 import net.sf.jsqlparser.statement.select.SelectExpressionItem;
 import net.sf.jsqlparser.statement.select.SelectItem;
+import net.sf.jsqlparser.statement.select.SelectVisitorAdapter;
+import net.sf.jsqlparser.statement.select.SubSelect;
 import org.springframework.util.CollectionUtils;
 
 /**
@@ -109,6 +115,28 @@ public class SqlParserSelectHelper {
         return (Select) statement;
     }
 
+    public static List<PlainSelect> getPlainSelects(PlainSelect plainSelect) {
+        List<PlainSelect> plainSelects = new ArrayList<>();
+        plainSelects.add(plainSelect);
+        plainSelect.accept(new SelectVisitorAdapter() {
+            @Override
+            public void visit(PlainSelect plainSelect) {
+                Expression whereExpression = plainSelect.getWhere();
+                if (whereExpression != null) {
+                    whereExpression.accept(new ExpressionVisitorAdapter() {
+                        @Override
+                        public void visit(SubSelect subSelect) {
+                            SelectBody subSelectBody = subSelect.getSelectBody();
+                            if (subSelectBody instanceof PlainSelect) {
+                                plainSelects.add((PlainSelect) subSelectBody);
+                            }
+                        }
+                    });
+                }
+            }
+        });
+        return plainSelects;
+    }
 
     public static List<String> getAllFields(String sql) {
 
@@ -244,14 +272,6 @@ public class SqlParserSelectHelper {
         return new ArrayList<>(result);
     }
 
-
-    public static boolean hasAggregateFunction(String sql) {
-        if (hasFunction(sql)) {
-            return true;
-        }
-        return hasGroupBy(sql);
-    }
-
     public static boolean hasGroupBy(String sql) {
         Select selectStatement = getSelect(sql);
         SelectBody selectBody = selectStatement.getSelectBody();
@@ -269,24 +289,36 @@ public class SqlParserSelectHelper {
         return false;
     }
 
-    public static boolean hasFunction(String sql) {
-        Select selectStatement = getSelect(sql);
-        SelectBody selectBody = selectStatement.getSelectBody();
+    public static boolean isLogicExpression(Expression whereExpression) {
+        return whereExpression instanceof AndExpression || (whereExpression instanceof OrExpression
+                || (whereExpression instanceof XorExpression));
+    }
 
-        if (!(selectBody instanceof PlainSelect)) {
-            return false;
+    public static String getColumnName(Expression leftExpression, Expression rightExpression) {
+        if (leftExpression instanceof Column) {
+            return ((Column) leftExpression).getColumnName();
         }
-        PlainSelect plainSelect = (PlainSelect) selectBody;
-        List<SelectItem> selectItems = plainSelect.getSelectItems();
-        AggregateFunctionVisitor visitor = new AggregateFunctionVisitor();
-        for (SelectItem selectItem : selectItems) {
-            selectItem.accept(visitor);
+        if (rightExpression instanceof Column) {
+            return ((Column) rightExpression).getColumnName();
         }
-        boolean selectFunction = visitor.hasAggregateFunction();
-        if (selectFunction) {
-            return true;
+        return "";
+    }
+
+    public static String getColumnName(Expression leftExpression) {
+        if (leftExpression instanceof Column) {
+            Column leftColumnName = (Column) leftExpression;
+            return leftColumnName.getColumnName();
         }
-        return false;
+        if (leftExpression instanceof Function) {
+            Function function = (Function) leftExpression;
+            if (!CollectionUtils.isEmpty(function.getParameters().getExpressions())) {
+                Expression expression = function.getParameters().getExpressions().get(0);
+                if (expression instanceof Column) {
+                    return ((Column) expression).getColumnName();
+                }
+            }
+        }
+        return "";
     }
 }
